requirements.txt
Django==1.5.3
South==0.8.1
-e git+https://github.com/rtyler/Spawning.git@f56c0f52330c0fb5c6a7322bc5b3a83f73863440#egg=Spawning-dev
Unidecode==0.04.13
Werkzeug==0.9.1
django-bootstrap-pagination==0.1.11
-e git+https://github.com/django-extensions/django-extensions.git@ad9dcbee9cce0aae40c09e8c253277650a4643e7#egg=django_extensions-dev
django-guardian==1.1.0.beta
django-pylibmc==0.5.0
djangorestframework==2.3.8
eventlet==0.12.1
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
greenlet==0.4.1
ipython==0.13.1
lxml==3.1.2
psycopg2==2.4.6
pylibmc==1.2.3
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
